Split classes into separate modules
